requirements.txt
asgiref==3.7.2
astroid==2.15.6
black==23.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.6
dill==0.3.7
Django==4.2.3
filelock==3.13.1
fsspec==2023.12.1
idna==3.6
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
llvmlite==0.41.1
MarkupSafe==2.1.3
mccabe==0.7.0
more-itertools==10.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
mysqlclient==2.2.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai-whisper==20231117
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pylint==2.17.5
regex==2023.10.3
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.5.2
tomli==2.0.1
tomlkit==0.12.1
torch==2.1.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.1.0
wrapt==1.15.0
